Add browser support excluding node modules import
