+++ package.json
@@ -4,6 +4,9 @@
   "description": "Advanced URL Connection String parser + generator.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
+  "browser": {
+    "./dist/inspect.js": "./dist/browser/inspect.js"
+  },
   "scripts": {
     "build": "tsc -p src && tsc -p test",
     "lint": "tslint --fix ./src/**/*.ts ./test/**/*.ts",
