+++ src/browser/inspect.ts
@@ -1,0 +1,3 @@
+export const setupCustomInspect = (obj: any) => {
+	// No custom inspect for browsers
+};
