+++ src/tsconfig.json
@@ -9,6 +9,6 @@
     "noUnusedLocals": true
   },
   "include": [
-    "./*.ts"
+    "./**/*.ts"
   ]
 }
